Projects: stub @aragon/api-react when not iframe
During development, it is sometimes convenient to load an app in the
main browser window, outside of the iframe.

However, this breaks the connection to @aragon/api-react, which counts
on a host page being available, and relays messages from that host into
the iframe.

This sets up a stubbed @aragon/api-react, which only gets loaded in if

1. NODE_ENV is not production
2. the page is not in an iframe

It currently implements very few of the features of @aragon/api-react,
but already enough to load the app and sign into GitHub.

To make it work:

* run `npm run start:dev` as usual
* wait until the app opens in the browser
* visit `localhost:3333` (instead of `localhost:3000`)